+++ src/frontend/apps/impress/src/features/docs/doc-management/api/useUpdateDocLink.tsx
@@ -1,4 +1,6 @@
+import { VariantType, useToastProvider } from '@openfun/cunningham-react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useTranslation } from 'react-i18next';
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
 import { Doc, KEY_DOC } from '@/docs/doc-management';
@@ -39,6 +41,8 @@ export function useUpdateDocLink({
 }: UpdateDocLinkProps = {}) {
   const queryClient = useQueryClient();
   const { broadcast } = useBroadcastStore();
+  const { toast } = useToastProvider();
+  const { t } = useTranslation();
 
   return useMutation<Doc, APIError, UpdateDocLinkParams>({
     mutationFn: updateDocLink,
@@ -51,6 +55,14 @@ export function useUpdateDocLink({
 
       // Broadcast to every user connected to the document
       broadcast(`${KEY_DOC}-${variable.id}`);
+
+      toast(
+        t('The document visibility has been updated.'),
+        VariantType.SUCCESS,
+        {
+          duration: 2000,
+        },
+      );
 
       onSuccess?.(data);
     },